After discussions with Matthew Petroff and Ken Knoblauch finally decide to make color vision deficiency transformations in linear RGB (as opposed to gamma-corrected sRGB) by default. The previous behavior is still available via linear = FALSE.
When creating colors via rgb() in simulate_cvd() they are rounded to
integers before the call because rgb() implicitly takes the floor() instead.
This will lead to minor differences in the output for linear = FALSE
compared to previous versions (for some colors).

Additional tests added that make sure that the different sequences of
steps for color vision deficiency simulation lead to identical() output.

Reference output for examples and cvd.R test updated to reflect the
changed defaults.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/colorspace@643 edb9625f-4e0d-4859-8d74-9fd3b1da38cb